requirements.txt
pymysql
elasticsearch
python-dotenv